+++ src/pygnssutils/gnssstreamer_cli.py
@@ -37,7 +37,6 @@ from time import sleep
 from types import FunctionType
 
 from pynmeagps import SocketWrapper
-from pysbf2 import SBFReader
 from pyubx2 import ERR_LOG, SETPOLL, UBXReader
 from pyubxutils.ubxsimulator import UBXSimulator
 from serial import Serial, SerialException
@@ -476,9 +475,9 @@ def main():
     ap.add_argument(
         "--protfilter",
         required=False,
-        help="1 = NMEA, 2 = UBX, 4 = RTCM3, 8 = SBF (can be OR'd; UBX and SBF are mutually exclusive, UBX takes precedence over SBF)",
+        help="1 = NMEA, 2 = UBX, 4 = RTCM3, 8 = SBF, 16 = QGC (can be OR'd)",
         type=int,
-        default=7,
+        default=31,
     )
     ap.add_argument(
         "--msgfilter",